Updated recommendation comparison
this function now does a lot less: it used to compute precision and recall, now only recall is computed. This means that we don't have to calculate similarities for all songs in the database, but only for the songs that are recommended by lastfm. The most I've seen recommended for one song is about 100, nowhere near 1000000.